backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pymupdf4llm
google-genai
openai
